+++ tests/test_division.py
@@ -31,6 +31,30 @@ class TestDivision(unittest.TestCase):
         self.assertEqual(list(division(a, b)),
                          [0.5, 1.0, 1.5])
 
+        a = np.array([1., 2., 3.])
+        b = 2.
+        assert isinstance(division(a, b), np.ndarray)
+        self.assertEqual(list(division(a, b)),
+                         [0.5, 1.0, 1.5])
+
+        a = 1.
+        b = (1., 2., 4.)
+        assert isinstance(division(a, b), tuple)
+        self.assertEqual(list(division(a, b)),
+                         [1.0, 0.5, 0.25])
+
+        a = 1.
+        b = [1., 2., 4.]
+        assert isinstance(division(a, b), list)
+        self.assertEqual(list(division(a, b)),
+                         [1.0, 0.5, 0.25])
+
+        a = 1.
+        b = np.array([1., 2., 4.])
+        assert isinstance(division(a, b), np.ndarray)
+        self.assertEqual(list(division(a, b)),
+                         [1.0, 0.5, 0.25])
+
         a = [1., 1., 1.]
         b = [2., 1., 0.]
         dd = division(a, b)
@@ -59,6 +83,13 @@ class TestDivision(unittest.TestCase):
         self.assertEqual(list(dd[0:2]), [0.5, 1.0])
         assert np.isnan(dd[2])
 
+        mask = [0, 0, 1]
+        a = np.ma.array([1., 1., 1.], mask=mask)
+        b = np.array([2., 1., 0.])
+        dd = division(a, b)
+        assert isinstance(dd, np.ma.masked_array)
+        self.assertEqual(list(dd), [0.5, 1.0, np.ma.masked])
+
         a = np.array([1., 1., 1.])
         mask = [0, 0, 1]
         b = np.ma.array([2., 1., 0.], mask=mask)
@@ -81,6 +112,30 @@ class TestDivision(unittest.TestCase):
         assert isinstance(div(a, b), list)
         self.assertEqual(list(div(a, b)),
                          [0.5, 1.0, 1.5])
+
+        a = np.array([1., 2., 3.])
+        b = 2.
+        assert isinstance(div(a, b), np.ndarray)
+        self.assertEqual(list(div(a, b)),
+                         [0.5, 1.0, 1.5])
+
+        a = 1.
+        b = (1., 2., 4.)
+        assert isinstance(div(a, b), tuple)
+        self.assertEqual(list(div(a, b)),
+                         [1.0, 0.5, 0.25])
+
+        a = 1.
+        b = [1., 2., 4.]
+        assert isinstance(div(a, b), list)
+        self.assertEqual(list(div(a, b)),
+                         [1.0, 0.5, 0.25])
+
+        a = 1.
+        b = np.array([1., 2., 4.])
+        assert isinstance(div(a, b), np.ndarray)
+        self.assertEqual(list(div(a, b)),
+                         [1.0, 0.5, 0.25])
 
         a = [1., 1., 1.]
         b = [2., 1., 0.]
@@ -109,6 +164,13 @@ class TestDivision(unittest.TestCase):
         assert isinstance(dd, np.ndarray)
         self.assertEqual(list(dd[0:2]), [0.5, 1.0])
         assert np.isnan(dd[2])
+
+        mask = [0, 0, 1]
+        a = np.ma.array([1., 1., 1.], mask=mask)
+        b = np.array([2., 1., 0.])
+        dd = div(a, b)
+        assert isinstance(dd, np.ma.masked_array)
+        self.assertEqual(list(dd), [0.5, 1.0, np.ma.masked])
 
         a = np.array([1., 1., 1.])
         mask = [0, 0, 1]